New module
Created a new module to refactor the Awsecrets module and make it
shorter (before was failing due tests with rubocop).
Also fixed some types in the spec.
